chips: rp2040: move the PL022 SPI driver into a shared rp2xxx crate
The RP2040 and the RP2350 fit the same Arm PL022 PrimeCell, with the same
register layout at different base addresses. Adding SPI to the rp2350 crate
would have meant a second copy of this file differing in three lines, so put
the driver in a new `rp2xxx` crate that both chips can depend on, following the
`nrf5x` and `stm32f4xx` precedent.

Only `spi` moves. `clocks`, `gpio` and `uart` each diverge by hundreds of lines
between the two chips and stay where they are.

The driver had exactly two chip-specific couplings, and both parameterize:

  - `clocks.get_frequency(Clock::Peripheral)` in `set_rate` and `get_rate`.
    The two `Clock` enums are unrelated types -- the RP2350 inserts `Hstx` --
    but the driver only ever names `Peripheral`, so `rp2xxx` defines a
    `PeripheralClock` trait and each chip's `Clocks` implements it.

  - `ChipSelectPolar<'a, gpio::RPGpioPin<'a>>`. `ChipSelectPolar` is already
    generic over `P: hil::gpio::Output`, so this becomes a type parameter.

Base addresses stay in the chip crate, which is where an address map belongs.
`rp2040::spi` is now a shim: the two `StaticRef`s, a `Spi<'a>` type alias with
the generics filled in, and `new_spi0`/`new_spi1`. Boards continue to name
`rp2040::spi::Spi` unchanged.

No functional change. On all four RP2040 boards -- raspberry_pi_pico,
pico_explorer_base, nano_rp2040_connect, raspberry_pi_pico_w -- `text`, `data`
and `bss` are byte-identical to the parent commit. Comparing symbols on
raspberry_pi_pico: all 592 keep their exact size, the sum of symbol sizes is
unchanged at 94274, and 581 keep their name. The 11 renamed ones pair one to
one by size and differ only in the SPI type's path, e.g.

    rp2040::spi::Spi::set_rate                                    216 bytes
    rp2xxx::spi::Spi<rp2040::clocks::Clocks,
                     rp2040::gpio::RPGpioPin>::set_rate           216 bytes

The binaries are not bit-identical: the linker places functions at different
addresses once the code lives in another crate, and one `SubSliceMut<u8>`
accessor is outlined instead of another, same size. Nothing else moved.
