+++ chips/rp2xxx/Cargo.toml
@@ -1,18 +1,15 @@
 # Licensed under the Apache License, Version 2.0 or the MIT License.
 # SPDX-License-Identifier: Apache-2.0 OR MIT
-# Copyright Tock Contributors 2022.
+# Copyright Tock Contributors 2026.
 
 [package]
-name = "rp2040"
+name = "rp2xxx"
 version.workspace = true
 authors.workspace = true
 edition.workspace = true
 
 [dependencies]
-cortexm0p = { path="../../arch/cortex-m0p" }
 kernel = { path = "../../kernel" }
-rp2xxx = { path = "../rp2xxx" }
-enum_primitive = { path = "../../libraries/enum_primitive" }
 
 [lints]
 workspace = true
